scripts/utah-trip-engine/requirements.txt
sqlalchemy>=2.0,<3.0
geoalchemy2>=0.14,<1.0
psycopg2-binary>=2.9,<3.0
pydantic>=2.0,<3.0
pydantic-settings>=2.0,<3.0
alembic>=1.13,<2.0
requests>=2.32,<3.0
beautifulsoup4>=4.12,<5.0
lxml>=5.0,<6.0
playwright>=1.45,<2.0
rapidfuzz>=3.9,<4.0
anthropic>=0.40,<1.0
typer>=0.12,<1.0
python-dotenv>=1.0,<2.0
shapely>=2.0,<3.0
tenacity>=8.5,<10.0
fastapi>=0.115,<1.0
uvicorn[standard]>=0.30,<1.0
jinja2>=3.1,<4.0
astral>=3.2,<4.0
pyyaml>=6.0,<7.0
